+++ src/tests/remotebase.tests.ts
@@ -3,6 +3,7 @@ import * as assert from 'assert';
 import {RemoteBase} from '../remotes/remotebase';
 import {Settings} from '../settings';
 import {Z80RegistersClass, Z80Registers, Z80RegistersStandardDecoder} from '../remotes/z80registers';
+import {Opcodes, Opcode} from '../disassembler/opcode';
 
 
 suite('RemoteBase', () => {
@@ -15,6 +16,8 @@ suite('RemoteBase', () => {
 		Settings.Init(cfg, '');
 		Z80RegistersClass.createRegisters();
 		Z80Registers.decoder=new Z80RegistersStandardDecoder();
+		// Restore 'rst 8' opcode
+		Opcodes[0xCF]=new Opcode(0xCF, "RST %s");
 	});
 
 
@@ -124,7 +127,7 @@ suite('RemoteBase', () => {
 			assert.equal(0x4567, bp2);
 		});
 
-		test('RST and step over', async () => {
+		test('RST (except 08) and step over', async () => {
 			const remote=new RemoteBaseMock();
 			const rem=remote as any;
 
@@ -137,7 +140,49 @@ suite('RemoteBase', () => {
 			assert.equal(undefined, bp2);
 		});
 
-		test('RST and step into', async () => {
+		test('RST 08 and step over', async () => {
+			const remote=new RemoteBaseMock();
+			const rem=remote as any;
+
+			remote.pc=0x8000;
+			remote.pcMemory[0]=0xCF;	// RST 8
+
+			let [opcode, bp1, bp2]=await rem.calcStepBp(true);	// stepOver
+			assert.equal(0xCF, opcode.code);
+			assert.equal(0x8001, bp1);
+			assert.equal(0x8002, bp2);
+		});
+
+		test('ESXDOS RST 08 and step over', async () => {
+			// Works the same as "RST 08 and step over"
+			// Extend 'rst 8' opcode for esxdos
+			Opcodes[0xCF].appendToOpcode(",#n");
+			const remote=new RemoteBaseMock();
+			const rem=remote as any;
+
+			remote.pc=0x8000;
+			remote.pcMemory[0]=0xCF;	// RST 8
+
+			let [opcode, bp1, bp2]=await rem.calcStepBp(true);	// stepOver
+			assert.equal(0xCF, opcode.code);
+			assert.equal(0x8001, bp1);
+			assert.equal(0x8002, bp2);
+		});
+
+		test('RST (except 08) and step into', async () => {
+			const remote=new RemoteBaseMock();
+			const rem=remote as any;
+
+			remote.pc=0x8000;
+			remote.pcMemory[0]=0xD7;	// RST 16
+
+			let [opcode, bp1, bp2]=await rem.calcStepBp(false);	// stepInto
+			assert.equal(0xD7, opcode.code);
+			assert.equal(0x0010, bp1);
+			assert.equal(undefined, bp2);
+		});
+
+		test('RST 08 and step into', async () => {
 			const remote=new RemoteBaseMock();
 			const rem=remote as any;
 
@@ -146,8 +191,24 @@ suite('RemoteBase', () => {
 
 			let [opcode, bp1, bp2]=await rem.calcStepBp(false);	// stepInto
 			assert.equal(0xCF, opcode.code);
-			assert.equal(0x0008, bp1);
-			assert.equal(undefined, bp2);
+			assert.equal(0x8001, bp1);
+			assert.equal(0x0008, bp2);
+		});
+
+		test('ESXDOS RST 08 and step into', async () => {
+			// Extend 'rst 8' opcode for esxdos
+			Opcodes[0xCF].appendToOpcode(",#n");
+			Settings.launch.disassemblerArgs.esxdosRst=true;
+			const remote=new RemoteBaseMock();
+			const rem=remote as any;
+
+			remote.pc=0x8000;
+			remote.pcMemory[0]=0xCF;	// RST 8
+
+			let [opcode, bp1, bp2]=await rem.calcStepBp(false);	// stepInto
+			assert.equal(0xCF, opcode.code);
+			assert.equal(0x8002, bp1);
+			assert.equal(0x0008, bp2);
 		});
 
 		test('Unconditional branches (JP, JR)', async () => {